+++ modules/cmsadmin/components/Menu.php
@@ -19,12 +19,12 @@ class Menu
 
     public function setCatByRewrite($catRewrite)
     {
-        $this->cat = (new \yii\db\Query())->select(['id', 'default_nav_id'])->from('cms_cat')->where(['rewrite' => $catRewrite])->one();
+        $this->cat = (new \yii\db\Query())->select(['id'])->from('cms_cat')->where(['rewrite' => $catRewrite])->one();
     }
 
     public function setCatById($catId)
     {
-        $this->cat = (new \yii\db\Query())->select(['id', 'default_nav_id'])->from('cms_cat')->where(['id' => $catId])->one();
+        $this->cat = (new \yii\db\Query())->select(['id'])->from('cms_cat')->where(['id' => $catId])->one();
     }
 
     public function setLangByShortCode($langShortCode)