+++ modules/cmsadmin/migrations/m141203_143052_cms_cat.php
@@ -10,16 +10,16 @@ class m141203_143052_cms_cat extends Migration
             'id' => 'pk',
             'name' => 'VARCHAR(180) NOT NULL',
             'rewrite' => 'VARCHAR(80) NOT NULL',
-            'default_nav_id' => 'INT(11) NOT NULL',
-            'is_default' => 'TINYINT(1) NOT NULL DEFAULT 0',
+            'default_nav_id' => 'INT(11) NOT NULL', // dropped in m151028_085932_add_is_home_in_nav
+            'is_default' => 'TINYINT(1) NOT NULL DEFAULT 0', // dropped in m151028_085932_add_is_home_in_nav
             'is_deleted' => 'TINYINT(1) NOT NULL default 0',
         ]);
 
         $this->insert('cms_cat', [
             'name' => 'Hauptnavigation',
             'rewrite' => 'default',
-            'default_nav_id' => 1,
-            'is_default' => 1,
+            //'default_nav_id' => 1,
+            //'is_default' => 1,
         ]);
     }